Disable typeahead matcher (task #2429)
Typeahead matcher is a bit of code which checks the records for
whether or not they match the search (typeahead) query/criteria.

Since we do all the filtering on the server and return the query
based results with AJAX, there is no need for this functionality.

If this is not enabled, then typeahead doesn't work correctly in
some scenarios.  One particular example is when we search for
something, which is NOT part of the display field, but part of
the configured other typeahead fields.  The server will return
correct results, which the matcher will further filter, removing
items which don't match by the display field, resulting in the
empty or incomplete suggestions list.

Additional confusion was caused by mismatch in the bootstrap
typeahead JavaScript libraries.  The relevant documentation is
here:

* http://plugins.upbootstrap.com/bootstrap-ajax-typeahead/
* https://github.com/biggora/bootstrap-ajax-typeahead

The NOT relevant documentation is here:

* https://github.com/twitter/typeahead.js
* ... and in most StackOverflow/Google search results.

Notice the version of the library in use by us (v0.0.5), and the
version for which most of the search results are (v0.11.1 or so).