Added CsvTrait unit test (task #2441)
Testing traits is no fun at all.  See:

* http://stackoverflow.com/a/31124840/151647
*
https://phpunit.de/manual/current/en/test-doubles.html#test-doubles.mocking-traits-and-abstract-classes

In the case of CsvTrait everything becomes even
more complicated since both methods that it provides
are protected (not public, cannot be called from
outside).  And testing any of the classes that
use CsvTrait is even more complicated.

So, for the migration period from our CSV parsing
implementation to the league/csv implementation, both
CsvTrait methods were copied verbatim into the test
class.  New implementations are there also.

This allowed to verify that the returned results from
old ways and new ways are identical, with one exception
of default values for returned records.

In the old implementation, we were making sure that all
the fields are returned, whether or not they are defined
in the CSV.  While doing so, we also set the missing fields
to default values.

League/CSV implementation provides a way to ensure that
all fields are returned, but their values will be set to
null where missing.

This shouldn't be a problem.  We have 3 use cases for the
CSV parsing currently: migrations, lists, and views.  All
of them, AFAIK, will do fine with null values instead of
empty strings, etc.  This will be verified at a later stage.