Removed DetailViewParser and IndexViewParser (task #2441)
DetailViewParser and IndexViewParser can't be helpful at this
stage, since different views can have different structure of the
CSV files and result in different rendering (consider the
difference of rendering an index view vs. a panel-based edit
view for example).

These two parsers will be replaced with a more generic ViewParser,
which will simply return a list of all records.  How to validate
and process them, is the job of the caller.  At least at this
stage.