Fix 'Permission denied' error (task #2477)
When opening the files for parsing only, the 'r' mode should be
used, as files might be in places, where web server cannot write.

`league/csv` library opens them by default with mode 'r+', which
makes sense for it, as it can both read and write CSV files and
not necessarily in the web context.