Mark related record links with label-primary
When related record links are rendered, it sometimes difficult to
distinguish them from external links.  For example, when display
field of a related record is set to something like email, and other
email fields are rendered as mailto: links, it is very difficult to
distinguish the link functionality - when clicked, will it open an
email client or will it redirect the user to the related record?

This commit adds the label-primary class on related record links to
make them appear slightly different.  With this, mailto: links will
be rendered as regular links, while related record email addresses
will be rendered as labeled links, making them standout a bit more.