Disable escaping of rendered values in associated record listings
Until now, lists of associated records were rendered with the
escaped values (sometimes even with double escaped values).  This
was breaking the fields which contained HTML.  For example, the
mailto: links.  This commit removes the escaping for now.

We'll need a better way to safeguard the value rendering, but at
least now the broken output of associated records is fixed.