API authentication handling (task #2629)
Since now we are extending from Application's AppController, the
authentication logic was slightly modified. We now check to see
if the AuthComponent is already loaded, and if it is, we overwrite
its configuration with our API's auth config. If the component is
not loaded, we just load it up and pass to it the API auth config.

Next we check if the API auth flag is disabled. In this case we
allow access to all API methods. This way we can retrieve current
user information, if API token was provided, which will allow
other access control systems to work as expected.