More stability to BaseFieldHandler
`BaseFieldHandler` always assumed that the `$table` parameter
is always an instance of the Table class, when generating field
names.  This might not always be true and makes things somewhat
more difficult to test.

This fix adds a bit more stability to the method, checking if the
`$table` is in fact an object.  Also, adds support for strings
and empty values.