More stability to Date, Time, and Datetime field handlers
Date, Time, and Datetime field handlers always assumed that
`$options['fieldDefinitions']` is present and is an instance
of the CsvField/DbField class.  Field definitions were used
to figure out if the input being rendered is for the required
field or not.

This fix adds the check for the assumption, and sets the
required field to false by default.