Added Validator warnings for related Documents (task #2913)
Until we sort out Documents and Files properly, this should save
from a few embarrasing occurances of `related(Documents)` instead
of `files(Documents)` in migration.csv files.

The first one provides the regular one-to-many relationship to the
Documents module, much like any other.  The second one uses the
Documents module as a container for the uploaded files.

While both are legitimate cases, the files one is needed much more
frequently.  So now a warning is issued for the regular related
usage.

The only case we have where we know for sure that the related is
needed is in the Files module, which is a join table for the
many-to-many relationship between Documents and FileStorage.  This
special case is avoided and the warning will not be printed out.