Bugfix for DbLists and DbListItems Controllers (task #2942)
Currently when trying to access DbLists controller you get the
following error:
Path does not exist [../config/CsvMigrations/views/Dblists/index.csv]

The reason this is happening is because it extends CsvMigrations
AppController which brings in a lot of functionality not supported
by DbLists. To fix this we extend the Application's AppController
instead.