Smarter way for field handlers to get value from data
Currently, we have multiple scenarios of passing the field
value to the field handler (`renderValue()`, etc):

* Using `$data` parameter
* Using `$options['entity']` parameter

On top of that, both of these can be either one of the following:

* An Entity instance
* An array from $this->request->data()
* An actual value of the field

In order to simplify the handling of these cases and make them
more consisten, we now do the following:

1. Try to rely on `$data` parameter in all cases.
2. Pass `$data` to `_getFieldValueFromData()` method, which checks
   the type of the variable to be either an instance of Entity,
   and instance of Request, or something else.
3. If the `$data` is empty, then we try to get the value from the
   `$options['entitiy']` if possible.

The last one should eventually be factored out and disappear.  It's
only there because we didn't do the clean data handling in the first
place.