Added Upgrade Shell (task #3206)
Upgrade Shell takes care of moving files from the old locations
to the new ones.  Run as follows:

```
$ bin/cake upgrade config/CsvMigrations/
```

This will reorganize files into config/Modules.  The shell DOES
NOT work with git, so you will be able to review the changes before
committing.

Plugin configurations, validations, and other logic will be updated
in the separate commits.  Also, after the upgrade, you'll need to
adjust the list names, as they are all moved into the Common module
for now.