Fixing unit tests
Now that we require a table name and field name for the field
handler constructor, we need some testing data.  A Fields module
was added, with the migration.csv file containing a field of each
type.

This doesn't fix all the tests yet, but it gets rid of about 140
errors in PHPUnit run.