Removed $moduleName parameter from getFieldsDefinitions()
Since getFieldsDefinitions() is a public non-static method, it
already knows how to get the module/table name, so the parameter
is unnecessary.

Furthermore, the behavior of the method was wrong when the parameter
was specified, as it was returning the current module's field
definitions if they were already set, before checking for the name
of the modules.

Together with this change, referencing calls were adjusted.  Also,
some unit tests improvements were introduced, moving, among other
things, MigrationTraitTest units into the FooTable tests.