Added support for stub fields in getFieldsDefinitions() method
Until now, getFieldsDefinitions() method was returning only the
real field definitions found in the CSV configuration.  This works
well except for cases where dummy fields are needed (combined fields,
virtual fields, etc).

With this addition of the stub fields, it is now possible to
specify which fields must be returned for sure, and to also
provide some defaults for those fields, in case they are not
found in the CSV configuration.