Make getFieldsDefinitions() much more resilient
getFieldsDefinitions() method is used in a lot of different
situations, so it helps if it tries to work through, rather than
bail early.  Now we try really hard to return the field defintiions,
if it's an empty list.

It's not the job of the MigrationTrait to figure out when empty
definitions are OK or not.  This should be done on a case-by-case
basis from those places that call this functionality.