Remove leftovers of BootstrapUI and QoboAdminPanel
QoboAdminPanel has been replaced with AdminLte, so we don't
need any references to it anymore.

BootstrapUI (from Friends of Cake) is not needed anymore either.

The leftovers started causing test runs to fail, due to something
in the recent upgrade of CakePHP 3.3.13.  With the previous version
it was running just fine, but with the latest the following error
appeared:

```
1) CsvMigrations\Test\TestCase\FieldHandlers\BooleanFieldHandlerTest::testGetSearchOptions
Undefined index: input

/home/travis/build/QoboLtd/cakephp-csv-migrations/vendor/friendsofcake/bootstrap-ui/src/View/Helper/FormHelper.php:350
/home/travis/build/QoboLtd/cakephp-csv-migrations/vendor/cakephp/cakephp/src/View/Helper/FormHelper.php:1063
/home/travis/build/QoboLtd/cakephp-csv-migrations/vendor/friendsofcake/bootstrap-ui/src/View/Helper/FormHelper.php:249
/home/travis/build/QoboLtd/cakephp-csv-migrations/src/FieldHandlers/BaseFieldHandler.php:306
/home/travis/build/QoboLtd/cakephp-csv-migrations/src/FieldHandlers/BooleanFieldHandler.php:112
/home/travis/build/QoboLtd/cakephp-csv-migrations/tests/TestCase/FieldHandlers/BooleanFieldHandlerTest.php:48
```

Once the leftovers are cleared out, the tests seem to be running fine
and there are no noticeable issues in the user interface.