Lookup results sorting (task #3285)
Sorting of lookup results is now applied on the actual query before
is run. The sorting is applied by the module's typeahead fields,
and if there is a parent module then its display field takes
precedence over the module's typeahead fields.

For example if a module has typeahead fields ['name', 'type'] then
these will be used for sorting the results. If it has a parent module
with display field 'title', then the sorting order will change to
['title', 'name', 'type'].