Files restructure (task #3299)
When associated files are added to the entity object they are stored
in a property matching the association name, for example if you have
a 'photos' field associated with the File Storage table, then the
associated files will be added to the entity under a property named
'photos_file_storage_file_storage'. With the file restructure logic
the associated files are moved under the field name (in this case
'photos') and the association named property is removed.