Panel expression for null values
Isset function will return false for keys that are part of the array
but their value is set to null. This is not the expected behavior.
Instead we check if the key is set in the array and then we add it
to the result no matter what is the value of it.