Make _getHandler() use passed options
In some cases, the user of field handler factory might want to
have more control over which field handler is being used.  For
example, HasManyFieldHandler is mostly used in tabs of related
data, and is handled from the view.

For that we've already supported the `$options['fieldDefinitions']`
in the handlers, but the factory was completely ignoring this, and
therefore calling the field definitions with stub fields, which
would result in default field types (String).

This commit fixes the issue, making `$options['fieldDefinitions']`
more powerful, allowingn the overwrite of the $field parameter with
prepared field definitions.