Switched `fieldToDb()` method to be static
`fieldToDb()` method is used in several different scenarios.  In
some, it's similar to other field handler methods, where we know
the table, field, options, etc.

But in some others (for example, during the migrations run), we
might not know or even have a table.  Since the method itself does
not rely on any runtime functionality, it can be made static and
used for the migration or testing purposes without the full blown
field handler intance.