Render element (task #3317)
Added new protected renderElement method that handles logic for
rendering appropriate element based on Field Handler class and
render method (input, value etc).

The method supports rendering
custom element by passing the element name using `$options['element']`
parameter. If the element does exist, it will be used, and the
Field Handler appropriate parameters will be passed to it.

If a custom element was not provided, then it will try and use the
specific Field Handler's render element. If there is no specific
render element for the Field Handler, it will use the Base Field
Handler element. If that does not exist either, then an exception
will be thrown.