Select2 search on related field (task #3277)
Added support on select2 search for related fields. This is the
case where the module we are doing the lookup search has a display
field or a typeahead field which is a related one.

For example we are searching in `Articles` but the display field is
author_name. In this case now, we are executing the search logic
in the `Authors` module, using the query string passed to us by the
select2 search field, then we get the `UUIs` return from the `Authors`
search and we append them in the search query of the `Articles` module
using `IN` condition in the `WHERE` clause.