Refactoring of ValidateShell (task #3495)
Before, the validation shell was finding the list of all modules
and sending the list of all modules to each check.

Now, the validation shell finds the list of all modules, iterates
over the list and sends a single module to each check.

This will help with the two upcoming changes:

* Allow a module parameter to the validation shell to validate a
  particular module, not all of them.
* A gradual migration of checks from the validation shell to the
  JSON schema of each configuration file.