Removed sanitizeValue from FieldHandlerInterface and BaseFieldHandler (task #3809)
Value sanitization is non-trivial for a variety of reasons (when
should it happen - before or after rendering code? how to handle
complex values before and after rendering code? etc.).

So, from now on, renderers are responsible for sanitizing the values
both in parameters and in the results.  Each renderer knows better
what's sane and what's not, so it makes most sense to do it there.