Removed silencing of ModuleConfig exceptions (task #3809)
In the older days, we had to catch configuration finding and parsing
exceptions, as our validation was weak, and the files were often
missing for legitimate reasons.

Since ModuleConfig has moved to JSON schema validation of the files,
and also started providing default values for when the configuration
files are not there, we don't need to catch, log, and silence these
exceptions any more.

If ModuleConfig throws an exception, than it needs to be taken care of
by the developer.  It should aso be as obvious as possible.  With the
silencing of the exceptions, it's annoying and non-obvious to figure
out and troubleshooting configuration issues, without constantly
running the Validator Shell.