Use Utility::findDirs() for consistency
Both the MigrationTrait and ValidateShell now use the same
mechanism for finding configuration modules, which is provided
by the cakephp-utils v3.2.0 (`Utility::findDirs()`).

This decreases the code base slightly, but more importantly makes
sure that what is used by the application is the same as wht have
been validated by the shell.