Pass the default value through an event (task #3567)
At the level of field handlers we don't have (and don't want
to have) much information about how to handle dynamic default
values.  The best option is to pass the defaults from the file
into an event, and if anything was triggered, then use the
result of the event.

Event name: `CsvMigrations.FieldHandler.DefaultValue`.

This can be listened to on the application level do provide
dynamic values like dates and times (in user seleted time zone,
for example), or current user, or anything else really.