Moved PrettifyTrait to BaseViewListener
Previously, PrettifyTrait was needed, because `_prettify()` method
was used both from the `src/Controller/Api/AppController.php` and
from some events.  This is no longer a case, and hans't been for
a while now (`git log -S_prettify` points back to October 2016).

The problem with having the PrettifyTrait came up in one of the
recent unit test failures (EventTest), which is related to some
behavior change for ReflectionClass in PHP 7.1.  Tests fail in
PHP 5.6 and event PHP 7, but do work in PHP 7.1.  Here is an
example test run on TravisCI:

https://travis-ci.org/QoboLtd/cakephp-csv-migrations/builds/263264906

Hopefully, removing the PrettifyTrait altogether will get rid of
the problem with:

```
PHP Fatal error:  Cannot use Cake\ORM\Table as Table because the
name is already in use in /src/PrettifyTrait.php on line 5
```