Change return value of ModelAfterSaveListener
`sendCalendarReminder()` method had a few issues related to its
return value:

* It was only returning the result of the last attempted email
  sending.  This is unpredictable behavior with multiple emails.
* It was returning the value of whatever Email::send() returns,
  without checking or modifying it in any way.
* Returning `false` from event listeners can bring an unexpected
  side effect of stopped propagation.  For more information see:
  https://book.cakephp.org/3.0/en/core-libraries/events.html#stopping-events

The returned value was changed to always be an array.  If no
attempts were made due to setup/initialization issues, an empty
array will be returned.

If any attempts were made, then the array will contain each tried
email address as the key, and the result of the attempt as the
value.  This will usually be either an array response from the
Email::send() or the exception that was thrown.