Use Config table, field, view in BaseFieldHandler (task #5061)
Removed nearly identical functionality for setting table, field,
and view in BaseFieldHandler, and, instead, utilized those that
are set in the Config instance.

This minimizes the codebase, and makes things a bit more consistent,
especially when relying on the Config in BaseFieldHandler and also
passing the same Config to providers.

Minor adjustment to the unit test: previously null values were not
allowed for table parameter.  Now, they are allowed, as the Config
will just instantiate a dummy table instance.  However, when passing
table objects, they should be instances of Cake\ORM\Table.