Add Field Handler Setting class and empty list option
Before the last refactoring of the field handlers, there was an
EMPTY_OPTION_LABEL constant defined in the BaseListFieldHandler
class, which was super handy in a variety of situations.

The cosntant was removed and temporarily replaced with hardcoded
options in different providers.  This commit fixes the temporary
removal by returning the constant.  However, since the class
does not exist anymore, we introduce a new Setting class in the
Field Handlers namespace, which can be used for storing all
common constants.