Many-to-many link post data (task #5612)
Adjusted form post data on many-to-many link action to use the
association name as key, instead of the module name. Module name
was working fine before because of the assumption that the same
many to many modules can only be associated ones. This is no longer
the case, thus the association name is dynamically generated.