Allow importing non-csv fields to the database (task #6913)
When we import CSV files from `qobo/cakephp-search` we convert
`metric()` fields into a pair of fields amount/value. If these fields
appear in the CSV the system tries to create a CsvField object out of it.

It fails. In that case we fallback to `$schema`
and try storing each fields (aka _amount, or _value) as a normal db field.