Disallow unique on non-required fields via Validate Shell (task #7348)
Having a unique attribute on the non-required fields allows for the
creation of the first record with an empty value, and then fails to
create additional records, as empty value of the field is not unique
anymore.

Since these cases are difficult to troubleshoot and are not very
practical, we disallow them via the validation.